Restore ArrayWrapper around the broadcasted path's from_array call and document that it prevents dask from hashing the whole buffer to name the array, which for a memmap silently loads the entire file into memory
