+++ reproject/_array_utils.py
@@ -4,7 +4,7 @@ from dask_image.ndinterp import map_coordinates as dask_image_map_coordinates
 from dask_image.ndinterp import spline_filter
 from scipy.ndimage import spline_filter as scipy_spline_filter
 
-__all__ = ["map_coordinates", "dask_map_coordinates", "sample_array_edges"]
+__all__ = ["map_coordinates", "dask_map_coordinates", "sample_array_edges", "ArrayWrapper"]
 
 
 def find_chunk_shape(shape, max_chunk_size=None):
@@ -337,3 +337,25 @@ def pad_dask_array_to_grid(array, bounds, target_shape, target_chunks):
         if len(pieces) > 1:
             array = da.concatenate(pieces, axis=idim)
     return array
+
+
+class ArrayWrapper:
+    """
+    A minimal getitem-only wrapper hiding an array from dask's tokenizer.
+
+    Passing a Numpy array (in particular a memmap) directly to
+    ``da.from_array`` can make dask hash the whole buffer to compute the array
+    name, which silently loads the entire file into memory (see
+    https://github.com/dask/dask/issues/11850). Wrapping the array so that
+    dask can only access it through ``__getitem__``, combined with an explicit
+    ``name=``, guarantees the data is only ever read chunk by chunk.
+    """
+
+    def __init__(self, array):
+        self._array = array
+        self.ndim = array.ndim
+        self.shape = array.shape
+        self.dtype = array.dtype
+
+    def __getitem__(self, item):
+        return self._array[item]
